Rebase refactoring on main (post GA API merge)
- Reset to main (which is now PR #175)
- Re-apply: _ODataBase, _BatchBase, _QueryBuilderBase, _BatchContext Protocol,
  _operation_context in base, Self type annotation
- Re-export multipart helpers from _batch.py for test compatibility
- Update test_sql_parse.py patch target to _odata_base.urlparse
